@@ -15,7 +15,7 @@
         }
     ],
     "require": {
-        "php": "^8.0.2",
+        "php": "^8.1",
         "ext-ctype": "*",
         "ext-filter": "*",
         "ext-hash": "*",
@@ -23,33 +23,35 @@
         "ext-openssl": "*",
         "ext-session": "*",
         "ext-tokenizer": "*",
+        "composer-runtime-api": "^2.2",
         "brick/math": "^0.9.3|^0.10.2|^0.11",
         "doctrine/inflector": "^2.0.5",
         "dragonmantank/cron-expression": "^3.3.2",
         "egulias/email-validator": "^3.2.1|^4.0",
         "fruitcake/php-cors": "^1.2",
         "guzzlehttp/uri-template": "^1.0",
-        "laravel/serializable-closure": "^1.2.2",
+        "laravel/prompts": "^0.1.9",
+        "laravel/serializable-closure": "^1.3",
         "league/commonmark": "^2.2.1",
         "league/flysystem": "^3.8.0",
-        "monolog/monolog": "^2.0",
-        "nesbot/carbon": "^2.62.1",
+        "monolog/monolog": "^3.0",
+        "nesbot/carbon": "^2.67",
         "nunomaduro/termwind": "^1.13",
         "psr/container": "^1.1.1|^2.0.1",
         "psr/log": "^1.0|^2.0|^3.0",
         "psr/simple-cache": "^1.0|^2.0|^3.0",
         "ramsey/uuid": "^4.7",
-        "symfony/console": "^6.0.9",
-        "symfony/error-handler": "^6.0",
-        "symfony/finder": "^6.0",
-        "symfony/http-foundation": "^6.0",
-        "symfony/http-kernel": "^6.0",
-        "symfony/mailer": "^6.0",
-        "symfony/mime": "^6.0",
-        "symfony/process": "^6.0",
-        "symfony/routing": "^6.0",
-        "symfony/uid": "^6.0",
-        "symfony/var-dumper": "^6.0",
+        "symfony/console": "^6.2",
+        "symfony/error-handler": "^6.2",
+        "symfony/finder": "^6.2",
+        "symfony/http-foundation": "^6.4",
+        "symfony/http-kernel": "^6.2",
+        "symfony/mailer": "^6.2",
+        "symfony/mime": "^6.2",
+        "symfony/process": "^6.2",
+        "symfony/routing": "^6.2",
+        "symfony/uid": "^6.2",
+        "symfony/var-dumper": "^6.2",
         "tijsverkoyen/css-to-inline-styles": "^2.2.5",
         "vlucas/phpdotenv": "^5.4.1",
         "voku/portable-ascii": "^2.0"
@@ -101,26 +104,30 @@
         "league/flysystem-read-only": "^3.3",
         "league/flysystem-sftp-v3": "^3.0",
         "mockery/mockery": "^1.5.1",
-        "orchestra/testbench-core": "^7.24",
+        "nyholm/psr7": "^1.2",
+        "orchestra/testbench-core": "^8.18",
         "pda/pheanstalk": "^4.0",
-        "phpstan/phpdoc-parser": "^1.15",
         "phpstan/phpstan": "^1.4.7",
-        "phpunit/phpunit": "^9.5.8",
-        "predis/predis": "^1.1.9|^2.0.2",
-        "symfony/cache": "^6.0",
-        "symfony/http-client": "^6.0"
+        "phpunit/phpunit": "^10.0.7",
+        "predis/predis": "^2.0.2",
+        "symfony/cache": "^6.2",
+        "symfony/http-client": "^6.2.4",
+        "symfony/psr-http-message-bridge": "^2.0"
     },
     "provide": {
         "psr/container-implementation": "1.1|2.0",
         "psr/simple-cache-implementation": "1.0|2.0|3.0"
     },
     "conflict": {
+        "carbonphp/carbon-doctrine-types": ">=3.0",
+        "doctrine/dbal": ">=4.0",
         "tightenco/collect": "<5.5.33"
     },
     "autoload": {
         "files": [
             "src/Illuminate/Collections/helpers.php",
             "src/Illuminate/Events/functions.php",
+            "src/Illuminate/Filesystem/functions.php",
             "src/Illuminate/Foundation/helpers.php",
             "src/Illuminate/Support/helpers.php"
         ],
@@ -172,15 +179,15 @@
         "mockery/mockery": "Required to use mocking (^1.5.1).",
         "nyholm/psr7": "Required to use PSR-7 bridging features (^1.2).",
         "pda/pheanstalk": "Required to use the beanstalk queue driver (^4.0).",
-        "phpunit/phpunit": "Required to use assertions and run tests (^9.5.8).",
-        "predis/predis": "Required to use the predis connector (^1.1.9|^2.0.2).",
+        "phpunit/phpunit": "Required to use assertions and run tests (^9.5.8|^10.0.7).",
+        "predis/predis": "Required to use the predis connector (^2.0.2).",
         "psr/http-message": "Required to allow Storage::put to accept a StreamInterface (^1.0).",
         "pusher/pusher-php-server": "Required to use the Pusher broadcast driver (^6.0|^7.0).",
-        "symfony/cache": "Required to PSR-6 cache bridge (^6.0).",
-        "symfony/filesystem": "Required to enable support for relative symbolic links (^6.0).",
-        "symfony/http-client": "Required to enable support for the Symfony API mail transports (^6.0).",
-        "symfony/mailgun-mailer": "Required to enable support for the Mailgun mail transport (^6.0).",
-        "symfony/postmark-mailer": "Required to enable support for the Postmark mail transport (^6.0).",
+        "symfony/cache": "Required to PSR-6 cache bridge (^6.2).",
+        "symfony/filesystem": "Required to enable support for relative symbolic links (^6.2).",
+        "symfony/http-client": "Required to enable support for the Symfony API mail transports (^6.2).",
+        "symfony/mailgun-mailer": "Required to enable support for the Mailgun mail transport (^6.2).",
+        "symfony/postmark-mailer": "Required to enable support for the Postmark mail transport (^6.2).",
         "symfony/psr-http-message-bridge": "Required to use PSR-7 bridging features (^2.0)."
     },
     "config": {
