@@ -14,14 +14,14 @@
         }
     ],
     "require": {
-        "php": "^8.0.2",
+        "php": "^8.1",
         "ext-hash": "*",
-        "illuminate/collections": "^9.0",
-        "illuminate/contracts": "^9.0",
-        "illuminate/http": "^9.0",
-        "illuminate/macroable": "^9.0",
-        "illuminate/queue": "^9.0",
-        "illuminate/support": "^9.0"
+        "illuminate/collections": "^10.0",
+        "illuminate/contracts": "^10.0",
+        "illuminate/http": "^10.0",
+        "illuminate/macroable": "^10.0",
+        "illuminate/queue": "^10.0",
+        "illuminate/support": "^10.0"
     },
     "autoload": {
         "psr-4": {
@@ -30,13 +30,13 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "9.x-dev"
+            "dev-master": "10.x-dev"
         }
     },
     "suggest": {
-        "illuminate/console": "Required to use the auth:clear-resets command (^9.0).",
-        "illuminate/queue": "Required to fire login / logout events (^9.0).",
-        "illuminate/session": "Required to use the session based guard (^9.0)."
+        "illuminate/console": "Required to use the auth:clear-resets command (^10.0).",
+        "illuminate/queue": "Required to fire login / logout events (^10.0).",
+        "illuminate/session": "Required to use the session based guard (^10.0)."
     },
     "config": {
         "sort-packages": true
