@@ -14,11 +14,11 @@
         }
     ],
     "require": {
-        "php": "^8.0.2",
-        "illuminate/collections": "^9.0",
-        "illuminate/contracts": "^9.0",
-        "illuminate/pipeline": "^9.0",
-        "illuminate/support": "^9.0"
+        "php": "^8.1",
+        "illuminate/collections": "^10.0",
+        "illuminate/contracts": "^10.0",
+        "illuminate/pipeline": "^10.0",
+        "illuminate/support": "^10.0"
     },
     "autoload": {
         "psr-4": {
@@ -27,7 +27,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "9.x-dev"
+            "dev-master": "10.x-dev"
         }
     },
     "suggest": {
