@@ -14,10 +14,10 @@
         }
     ],
     "require": {
-        "php": "^8.0.2",
-        "illuminate/conditionable": "^9.0",
-        "illuminate/contracts": "^9.0",
-        "illuminate/macroable": "^9.0"
+        "php": "^8.1",
+        "illuminate/conditionable": "^10.0",
+        "illuminate/contracts": "^10.0",
+        "illuminate/macroable": "^10.0"
     },
     "autoload": {
         "psr-4": {
@@ -29,11 +29,11 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "9.x-dev"
+            "dev-master": "10.x-dev"
         }
     },
     "suggest": {
-        "symfony/var-dumper": "Required to use the dump method (^6.0)."
+        "symfony/var-dumper": "Required to use the dump method (^6.2)."
     },
     "config": {
         "sort-packages": true
