@@ -2,198 +2,104 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\CategoryRequest;
 use App\Models\Category;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Auth;
-use Throwable;
+use Yajra\DataTables\Facades\DataTables;
 
 class CategoryController extends Controller
 {
     /**
      * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
      */
-    public function index()
+    public function index(Request $request)
     {
-        if (!Auth::user()->can('category-list')) {
-            $response = array(
-                'message' => trans('no_permission_message')
-            );
-            return redirect(route('home'))->withErrors($response);
+        if ($request->ajax()) {
+            $data = Category::get();
+            return DataTables::of($data)
+                    ->addIndexColumn()
+                    ->addColumn('category', function($row){
+                            $name = $row->name ?? '';
+                            return $name;
+                    })
+                    ->addColumn('action', function($row){
+                            $btn = '<a href="javascript:void(0)" class="edit btn btn-primary btn-sm">View</a>';
+                            $btn .= '<a href="'. route('category.edit', $row->id) .'" class="edit btn btn-primary btn-sm mx-2">Edit</a>';
+                            $btn .= '<a href="javascript:void(0)" onclick="deletecat('. $row->id .');" class="edit btn btn-primary btn-sm">Delete</a>';
+                            return $btn;
+                    })
+                    ->rawColumns(['action', 'category'])
+                    ->make(true);
         }
-        return view('category.index');
+
+        return view('category.list');
     }
 
+    /**
+     * Show the form for creating a new resource.
+     */
+    public function create()
+    {
+        return view('category.add');
+    }
 
     /**
      * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(CategoryRequest $request)
     {
-        if (!Auth::user()->can('category-create')) {
-            $response = array(
-                'error' => true,
-                'message' => trans('no_permission_message')
-            );
-            return response()->json($response);
-        }
-        $request->validate([
-            'name' => 'required',
-        ]);
-
         try {
-            $category = new Category();
-            $category->name = $request->name;
+            $category = new Category;
+            $category->name = $request->category;
             $category->save();
-            $response = [
-                'error' => false,
-                'message' => trans('data_store_successfully')
-            ];
-        } catch (Throwable $e) {
-            $response = array(
-                'error' => true,
-                'message' => trans('error_occurred'),
-                'data' => $e
-            );
+
+            return redirect()->route('category.index')->with('success', 'Category add successfully.');
+        } catch (\Throwable $th) {
+            //throw $th;
+            return redirect()->back()->with('error', 'Something went wrong!');
         }
-        return response()->json($response);
     }
 
     /**
      * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
      */
-    public function show()
+    public function show(Category $category)
     {
-        if (!Auth::user()->can('category-list')) {
-            $response = array(
-                'message' => trans('no_permission_message')
-            );
-            return response()->json($response);
-        }
-        $offset = 0;
-        $limit = 10;
-        $sort = 'id';
-        $order = 'DESC';
-
-        if (isset($_GET['offset']))
-            $offset = $_GET['offset'];
-        if (isset($_GET['limit']))
-            $limit = $_GET['limit'];
-
-        if (isset($_GET['sort']))
-            $sort = $_GET['sort'];
-        if (isset($_GET['order']))
-            $order = $_GET['order'];
-
-        $sql = Category::where('id', '!=', 0);
-        if (isset($_GET['search']) && !empty($_GET['search'])) {
-            $search = $_GET['search'];
-            $sql->where('id', 'LIKE', "%$search%")->orwhere('name', 'LIKE', "%$search%")->orwhere('status', 'LIKE', "%$search%");
-        }
-        $total = $sql->count();
-
-        $sql->orderBy($sort, $order)->skip($offset)->take($limit);
-        $res = $sql->get();
-
-        $bulkData = array();
-        $bulkData['total'] = $total;
-        $rows = array();
-        $tempRow = array();
-        $no = 1;
-        foreach ($res as $row) {
-            $operate = '<a class="btn btn-xs btn-gradient-primary btn-rounded btn-icon editdata" data-id=' . $row->id . ' data-url=' . url('category') . ' title="Edit" data-toggle="modal" data-target="#editModal"><i class="fa fa-edit"></i></a>&nbsp;&nbsp;';
-            $operate .= '<a class="btn btn-xs btn-gradient-danger btn-rounded btn-icon deletedata" data-id=' . $row->id . ' data-url=' . url('category', $row->id) . ' title="Delete"><i class="fa fa-trash"></i></a>';
-
-            $tempRow['id'] = $row->id;
-            $tempRow['no'] = $no++;
-            $tempRow['name'] = $row->name;
-            $tempRow['status'] = $row->status;
-            $tempRow['operate'] = $operate;
-            $rows[] = $tempRow;
-        }
-
-        $bulkData['rows'] = $rows;
-        return response()->json($bulkData);
+        //
     }
 
     /**
      * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
      */
-    public function edit($id)
+    public function edit(Category $category)
     {
-        $category = Category::find($id);
-        return response($category);
+
+        return view('category.edit', compact('category'));
     }
 
-    public function update(Request $request)
+    /**
+     * Update the specified resource in storage.
+     */
+    public function update(CategoryRequest $request, Category $category)
     {
-        if (!Auth::user()->can('category-edit')) {
-            $response = array(
-                'error' => true,
-                'message' => trans('no_permission_message')
-            );
-            return response()->json($response);
-        }
-        $request->validate([
-            'name' => 'required',
-        ]);
-
         try {
-            $category = Category::find($request->id);
-            $category->status = $request->status;
-            $category->name = $request->name;
-            $category->save();
-            $response = [
-                'error' => false,
-                'message' => trans('data_update_successfully'),
-            ];
-        } catch (Throwable $e) {
-            $response = array(
-                'error' => true,
-                'message' => trans('error_occurred'),
-                'data' => $e
-            );
-        }
-        return response()->json($response);
-    }
+            $category->name = $request->category;
+            $category->update();
 
+            return redirect()->route('category.index')->with('success', 'Category update successfully.');
+        } catch (\Throwable $th) {
+            return redirect()->back()->with('error', 'Something went wrong!');
+        }
+    }
 
     /**
      * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Category $category)
     {
-        if (!Auth::user()->can('category-delete')) {
-            $response = array(
-                'message' => trans('no_permission_message')
-            );
-            return response()->json($response);
-        }
-        try {
-            Category::find($id)->delete();
-            $response = [
-                'error' => false,
-                'message' => trans('data_delete_successfully')
-            ];
-        } catch (Throwable $e) {
-            $response = array(
-                'error' => true,
-                'message' => trans('error_occurred')
-            );
-        }
-        return response()->json($response);
+        $category->delete();
+        return response()->json([
+            'success' => 1,
+        ]);
     }
 }
