@@ -1,4 +1,4 @@
-{{-- @extends('layouts.app')
+@extends('layouts.app')
 
 @section('content')
 <div class="container">
@@ -44,110 +44,4 @@
         </div>
     </div>
 </div>
-@endsection --}}
-
-
-
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <!-- Required meta tags -->
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-
-    <title>{{ __('Reset Password') }} || {{ config('app.name') }}</title>
-
-    @include('layouts.include')
-
-</head>
-
-<body>
-    <div class="container-scroller">
-        <div class="container-fluid page-body-wrapper full-page-wrapper">
-            <div class="content-wrapper d-flex align-items-center auth">
-                <div class="row flex-grow">
-                    <div class="col-lg-4 mx-auto">
-                        <div class="auth-form-light text-left p-5">
-                            <div class="brand-logo text-center">
-                                {{-- <img src="{{asset('public/' . config('global.LOGO1')) }}" alt="logo"> --}}
-                                {{-- <img src="{{ asset('publiclogo.svg') }}" alt="logo"> --}}
-                                <img src="{{asset('publicstorage/'.env('LOGO1')) }}" alt="logo">
-                            </div>
-                            @if (session('status'))
-                                <div class="alert alert-success" role="alert">
-                                    {{ session('status') }}
-                                </div>
-                            @endif
-                            <form method="POST" action="{{ route('password.email') }}">
-                                @csrf
-                                <div class="form-group">
-                                    <label>{{ __('email') }}</label>
-                                    {{-- <input type="text" name="username" required class="form-control form-control-lg" placeholder="{{__('username')}}"> --}}
-                                    <input id="email" type="email" class="form-control form-control-lg" name="email"
-                                        value="{{ old('email') }}" required autocomplete="email" autofocus
-                                        placeholder="{{ __('email') }}">
-                                </div>
-                                <div class="mt-3">
-                                    <input type="submit" name="btnlogin" value="{{ __('Send Password Reset Link') }}"
-                                        class="btn btn-block btn-theme btn-lg font-weight-medium auth-form-btn" />
-                                </div>
-                            </form>
-                        </div>
-                    </div>
-                </div>
-            </div>
-            <!-- content-wrapper ends -->
-        </div>
-        <!-- page-body-wrapper ends -->
-    </div>
-
-    <script src="{{ asset('public/assets/js/vendor.bundle.base.js') }}"></script>
-    <script src="{{ asset('public/assets/js/jquery.validate.min.js') }}"></script>
-    <script src="{{ asset('public/assets/jquery-toast-plugin/jquery.toast.min.js') }}"></script>
-
-    <script type='text/javascript'>
-        $("#frmLogin").validate({
-            rules: {
-                username: "required",
-                password: "required",
-            },
-            errorPlacement: function(label, element) {
-                label.addClass('mt-2 text-danger');
-                label.insertAfter(element);
-            },
-            highlight: function(element, errorClass) {
-                $(element).parent().addClass('has-danger')
-                $(element).addClass('form-control-danger')
-            }
-        });
-    </script>
-</body>
-
-@if (Session::has('error'))
-    <script type='text/javascript'>
-        $.toast({
-            text: '{{ Session::get('error') }}',
-            showHideTransition: 'slide',
-            icon: 'error',
-            loaderBg: '#f2a654',
-            position: 'top-right'
-        });
-    </script>
-@endif
-
-@if ($errors->any())
-    @foreach ($errors->all() as $error)
-        <script type='text/javascript'>
-            $.toast({
-                text: '{{ $error }}',
-                showHideTransition: 'slide',
-                icon: 'error',
-                loaderBg: '#f2a654',
-                position: 'top-right'
-            });
-        </script>
-    @endforeach
-@endif
-
-</html>
+@endsection
